requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
psycopg2-binary
pydantic
google-generativeai
celery
redis